% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_build_request.R
\name{cr_build_request}
\alias{cr_build_request}
\title{Build a valid request object from the checked parameters}
\usage{
cr_build_request(params, verbose = FALSE, ds_id)
}
\value{
a httr2 request object
}
\description{
Build a valid request object from the checked parameters
}
