% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_check_params.R
\name{cr_check_params}
\alias{cr_check_params}
\title{Check that input parameters are valid and in compliance with COMEX APIs.}
\usage{
cr_check_params(
  freq,
  reporter,
  partner,
  stat_procedure,
  product,
  time,
  indicators,
  flow,
  update,
  verbose
)
}
\description{
Check that input parameters are valid and in compliance with COMEX APIs.
}
