% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_get_data.R
\name{cr_get_data}
\alias{cr_get_data}
\title{Get trade data from the Comex API}
\usage{
cr_get_data(
  freq,
  reporter,
  partner,
  product,
  time,
  indicators,
  flow,
  process = TRUE,
  tidy_cols = TRUE,
  verbose = FALSE,
  update = T,
  requests_per_second = 10/60,
  ds_id
)
}
\value{
A data.frame with trade data or, if \code{process = F}, a httr2 response object.
}
\description{
This function will get Comex data from the new API.
@param ds_id This will let you define which dataset you want to query.
Most likely you want: 045409, as it is the standard trade dataset. Check
\verb{https://ec.europa.eu/eurostat/api/comext/dissemination/sdmx/2.1/dataflow/ESTAT/all} for all available datasets.
}
