% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cr_get_ref_table}
\alias{cr_get_ref_table}
\title{Get reference table from package data}
\usage{
cr_get_ref_table(dataset_id, update = FALSE, verbose = FALSE)
}
\arguments{
\item{dataset_id}{The dataset ID, which is either partner, reporter or a valid classification scheme.}
}
\value{
a tidy dataset with a reference table
}
\description{
The first time, the function will read from disk, the second time from the environment. In the case of a necessary update the new data will be saved to the environment for the current session.
You can use this table to look at the reference tables and if necessary extract respective classification codes by hand. In general we would recommend the function \code{ct_commodity_lookup} for this purpose. It uses the present function in the backend.
}
