% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_perform_request.R
\name{cr_perform_request}
\alias{cr_perform_request}
\title{Performs the request to the Comex API}
\usage{
cr_perform_request(req, requests_per_second, verbose = FALSE)
}
\arguments{
\item{req}{a valid comtrade request built by the \code{cr_build_request()} function}
}
\description{
Performs the request to the Comex API
}
