% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_process_response.R
\name{cr_process_response}
\alias{cr_process_response}
\title{Processes the response object}
\usage{
cr_process_response(resp, verbose = FALSE, tidy_cols)
}
\arguments{
\item{resp}{a valid httr2 response object created from the function \code{cr_perform_request()}}
}
\value{
a data.frame object with the results
}
\description{
Processes the response object
}
